+++ milvus/client/check.py
@@ -12,7 +12,7 @@ def is_legal_host(host):
 
 
 def is_legal_port(port):
-    if isinstance(port, str):
+    if isinstance(port, (str, int)):
         try:
             _port = int(port)
         except ValueError:
@@ -25,10 +25,13 @@ def is_legal_port(port):
 
 
 def is_legal_uri(uri):
+    if uri is None:
+        return True
+
     try:
         _uri = urlparse(uri)
         return _uri.scheme == 'tcp'
-    except (AttributeError, ):
+    except (AttributeError, ValueError, TypeError):
         return False