+++ milvus/client/hooks/abstract.py
@@ -2,13 +2,18 @@ class AbstractHook:
     pass
 
 
-class BaseaSearchHook(AbstractHook):
+class BaseSearchHook(AbstractHook):
+    """
+    Base class of search hook. Inherited from hook class of search
+    """
 
-    def prepare_data(self, *args, **kwargs):
-        """
-        Prepare grpc data
-        """
-        pass
+    # def prepare_data(self, *args, **kwargs):
+    #     """
+    #     Prepare grpc data.
+    #
+    #     This is an EXPERIMENTAL API.
+    #     """
+    #     pass
 
     def pre_search(self, *args, **kwargs):
         """
@@ -29,7 +34,7 @@ class BaseaSearchHook(AbstractHook):
         return bool variables. False default
 
         When return True, the method tell caller to return raw response received from
-        server directly.
+        server directly. In this situation, method `handle_response` will not be invoked.
         """
         return False